+++ scilab/modules/fileio/src/c/mput.c
@@ -87,7 +87,7 @@ void C2F(mputnc) (int *fd, void * res, int *n1, char *type, int *ierr)
     switch ( type[0] )
     {
         case 'i' :
-            MPUT_GEN_NC(long, c1);
+            MPUT_GEN_NC(int, c1);
             break;
         case 'l' :
             MPUT_GEN_NC(long long, c1);
@@ -108,7 +108,7 @@ void C2F(mputnc) (int *fd, void * res, int *n1, char *type, int *ierr)
             switch ( c1 )
             {
                 case 'i' :
-                    MPUT_GEN_NC(unsigned long, c2);
+                    MPUT_GEN_NC(unsigned int, c2);
                     break;
                 case 'l' :
                     MPUT_GEN_NC(unsigned long long, c2);
@@ -189,7 +189,7 @@ void mput2 (FILE *fa, int swap2, double *res, int n, char *type, int *ierr)
     switch ( type[0] )
     {
         case 'i' :
-            MPUT_GEN(long, c1);
+            MPUT_GEN(int, c1);
             break;
         case 'l' :
             MPUT_GEN(long long, c1);
@@ -210,7 +210,7 @@ void mput2 (FILE *fa, int swap2, double *res, int n, char *type, int *ierr)
             switch ( c1 )
             {
                 case 'i' :
-                    MPUT_GEN(unsigned long, c2);
+                    MPUT_GEN(unsigned int, c2);
                     break;
                 case 'l' :
                     MPUT_GEN(unsigned long long, c2);
